requirements.txt
fastapi
uvicorn
python-multipart
requests
firebase-admin
twilio
langchain
langchain-community
faiss-cpu
sentence-transformers
groq
python-dotenv
pandas
openpyxl
pymupdf
jinja2
llmwhisperer-client
pdfplumber
camelot-py[cv]
PyPDF2
faster-whisper
langchain-groq
